feat(java): string concatenation
