+++ new/detector/implementation/java/string/string.go
@@ -6,6 +6,7 @@ import (
 	"github.com/bearer/bearer/pkg/commands/process/settings"
 	"github.com/bearer/bearer/pkg/util/stringutil"
 
+	"github.com/bearer/bearer/new/detector/implementation/generic"
 	generictypes "github.com/bearer/bearer/new/detector/implementation/generic/types"
 	languagetypes "github.com/bearer/bearer/new/language/types"
 )
@@ -27,11 +28,16 @@ func (detector *stringDetector) DetectAt(
 	ruleReferenceType settings.RuleReferenceScope,
 	evaluator types.Evaluator,
 ) ([]interface{}, error) {
-	if node.Type() == "string_literal" {
+	switch node.Type() {
+	case "string_literal":
 		return []interface{}{generictypes.String{
 			Value:     stringutil.StripQuotes(node.Content()),
 			IsLiteral: true,
 		}}, nil
+	case "binary_expression":
+		if node.AnonymousChild(0).Content() == "+" {
+			return generic.ConcatenateChildStrings(node, evaluator)
+		}
 	}
 
 	return nil, nil
