+++ new/detector/implementation/java/testdata/string.java
@@ -1,0 +1,8 @@
+public class Greet {
+  const Greeting = "Hello World";
+
+  public static void main(String[] args)
+  {
+    var s = Greeting + "!";
+  }
+}
