--- new/detector/implementation/java/testdata/string_literal.java
@@ -1,3 +1,0 @@
-public class Greet {
-  const Greeting = "Hello World";
-}
